pyproject.toml
[tool.poetry]
name = "gtt"
version = "0.1.0"
description = "Track your good habits"
authors = ["Nikita V. <BrutalCutie@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
ipython = "^8.30.0"
django = "^5.1.4"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
python-dotenv = "^1.0.1"
redis = "^5.2.1"
celery = "^5.4.0"
eventlet = "^0.38.1"
django-celery-beat = "^2.7.0"
django-cors-headers = "^4.6.0"
coverage = "^7.6.9"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
